Crate alarm_clock

Alarm clock program for ATMEGA328P microprocessor

Modules

clocks 🔒

Clocks, date and time management

inputs 🔒

Inputs, either of the environment or of the user

outputs 🔒

Outputs of the clock setup, for humans

serial_commands 🔒

Serial communication, with commands from the user

Structs

MainState 🔒

The main state of the whole program, updated at every loop, holding the memory.

Constants

ALARM_DAWN_DURATION_MINUTES 🔒

Duration of the dawn before the sunrise

ALARM_WEEKEND_SUNRISE 🔒

Time of the sunrise (alarm) during the week-end

ALARM_WEEK_SUNRISE 🔒

Time of the sunrise (alarm) during the week

BUTTON_LOGICAL_LEVEL_HIGH 🔒

Mapping between the electric levels (+3.3V or +5V) and the logical level of the button

BUZZER_LOGICAL_LEVEL_HIGH 🔒

Mapping between the electric levels (+3.3V or +5V) and the logical level of the buzzer

DCF77_LOGICAL_LEVEL_HIGH 🔒

Mapping between the electric levels (+3.3V or +5V) and the logical level of the DCF77 receiver

LED_STRIP_COUNT 🔒

Number of individual leds on the strip

LED_STRIP_MAX_INTENSITY 🔒

Value of the brightness of the simulated day

LUMINOSITY_LOGICAL_LEVEL_HIGH 🔒

Mapping between the electric levels (+3.3V or +5V) and the logical level of the luminosity sensor

PROXIMITY_LOGICAL_LEVEL_HIGH 🔒

Mapping between the electric levels (+3.3V or +5V) and the logical level of the proximity sensor

Functions

panic 🔒

Panic handler: do nothing

Type Definitions

ButtonInput 🔒

BuzzerOutput 🔒

Dcf77Input 🔒

DisplaySpiClkOutput 🔒

DisplaySpiCsOutput 🔒

DisplaySpiMosiOutput 🔒

LedStripOutput 🔒

LuminosityInput 🔒

ProximityInput 🔒