Module alarm_clock::clocks::dcf77

DCF77 decoder: take as input the bits polled from the receiver, and return a complete datetime.

Re-exports

pub use self::protocol::Dcf77SignalVariant;

Modules

phase_detector 🔒

Phase detector, to detect the beginning of a second in the polled bits.

polled_values 🔒

Structure and methods related to the polled values of the pin linked to the DCF77 receiver.

protocol 🔒

Decode the bits of the DCF77 signal

Structs

Dcf77

Dcf77 decoder main structure

Enums

WorkflowError

Errors that may arise during the decoding process

Constants

POLLED_SAMPLES_FREQUENCY