Struct alarm_clock::clocks::dcf77::Dcf77

pub struct Dcf77 {
    last_update_timer: Option<Timer>,
    polled_values: PolledValues,
    phase_detector: PhaseDetector,
    last_peak_update: Option<Timer>,
    signals: Signals,
}

Dcf77 decoder main structure

Fields
last_update_timer: Option<Timer>

Time of the last polled value

polled_values: PolledValues

Array holding the past boolean polled values of the DFC77 pin input at a “high” rate

phase_detector: PhaseDetector

Detector of the phase of the signal, that is, detect the beginning of UTC seconds among the polled values

last_peak_update: Option<Timer>

Time of the last detected peak of the correlation between the samples and the pattern, signaling a new bit

signals: Signals

Array holding the decoded, 1Hz-bits of the DCF77 signal

Implementations

impl Dcf77

pub fn run(&mut self) -> Result<Option<Datetime>, WorkflowError>

Main call

pub fn process_new_polled_values( &mut self, current_timer: Timer, polled_value: bool ) -> Result<Option<Dcf77SignalVariant>, WorkflowError>

Detect if the container of the polled value holds a new one; if so, process it, and if a new bit of the signal, return it.