Module alarm_clock::clocks::timer

Timer based on the internal clock, moderately accurate.

Structs

Timer

Timer structure, new-type pattern

TimerPolling 🔒

Timer structure holding polled values of the DCF77 input

Constants

MILLIS_INCREMENT

Number of milliseconds to increment the counter by at each interrupt. Possible Values:

POLLED_SAMPLES_PERIOD_MS

Downsampling factor for the polling of the DCF77 input

PRESCALER 🔒

Prescaler of the internal timer (see the doc of the microprocessor)

TIMER_COUNTS 🔒

Tick counts before raising an interrupt

TIMER_MAX 🔒

The downsampling factor is based on the modulo of the timer. The rolling of the timer must occur at the biggest multiple of the POLLED_SAMPLES_PERIOD_MS without overflowing the capacity of the integer type.

Statics

TIMER 🔒

Global timer object

Functions

get_polled_values

Get the polled value of DCF77, if any, with the time when it was set to discrimate between an old value already processed and a new one.

get_timer

Get the current value of the timer.

init

Initialize the registrers for the hardware timer.