Struct alarm_clock::inputs::BoolInput

pub struct BoolInput<PIN: PinOps> {
    pin: Pin<Input<PullUp>, PIN>,
    logical_level_high: bool,
    pub value: bool,
}

Generic boolean input.

Fields
pin: Pin<Input<PullUp>, PIN>

Pin to read the state from.

logical_level_high: bool

Mapping between the electric level of the pin (+3.3V or +5V) and the logical level of the input.

value: bool

Logical value of the input: true for an active state, false for an inactive state.

Implementations

impl<PIN: PinOps> BoolInput<PIN>

pub fn init(pin: Pin<impl Io, PIN>, logical_level_high: bool) -> Self

Initialize the structure.

pub fn update(&mut self)

Update the value by reading the electric state of the pin.