Struct alarm_clock::inputs::Inputs

pub struct Inputs {
    pub button: BoolInput<PD4>,
    pub luminosity: BoolInput<PD3>,
    pub proximity: BoolInput<PD6>,
}

Misc. inputs

Fields
button: BoolInput<PD4>

Button to be activated by the user

luminosity: BoolInput<PD3>

Ambiant luminosity sensor

proximity: BoolInput<PD6>

Infrared proximity/motion sensor

Implementations

impl Inputs

pub fn init( button_pin: Pin<impl Io, PD4>, button_elec_level_to_logical_level: bool, luminosity_pin: Pin<impl Io, PD3>, luminosity_elec_level_to_logical_level: bool, proximity_pin: Pin<impl Io, PD6>, proximity_elec_level_to_logical_level: bool ) -> Self

Initialize all the inputs, in particular with the pinout and the electrical to logical mapping.

pub fn update(&mut self)

Update the values of the inputs by reading the electric state of the pins.