Struct alarm_clock::outputs::display::Display

pub struct Display {
    display_interface: MAX7219<PinConnector<Pin<Output, PB3>, Pin<Output, PB2>, Pin<Output, PB5>>>,
    frame_buffer: FrameBuffer,
}

Main structure for the display

Fields
display_interface: MAX7219<PinConnector<Pin<Output, PB3>, Pin<Output, PB2>, Pin<Output, PB5>>>

Electronic interface with the display

frame_buffer: FrameBuffer

Frame buffer containing the pixels to display

Implementations

impl Display

pub fn init( data: Pin<impl Io, PB3>, cs: Pin<impl Io, PB2>, clk: Pin<impl Io, PB5> ) -> Self

Initialize the display.

pub fn clear_display(&mut self)

Clear the content of the display, leaving it blank

pub fn render(&mut self)

Actually render (publish) the content of the frame buffer.

pub fn set_intensity(&mut self, intensity: DisplayIntensity)

Set the intensity of the display.

pub fn set_at(&mut self, column_index: usize, value: &[u8])

Set one or more columns of the display begininng at the specified index with the content of the slice of u8.

pub fn write_time(&mut self, datetime: Option<Datetime>)

Write the time HH:MM at the beginning of the display.