Struct alarm_clock::outputs::display::FrameBuffer

struct FrameBuffer {
    current: ColumnMajorBuffer,
    buffer: ColumnMajorBuffer,
    column_index: usize,
    current_intensity: DisplayIntensity,
    buffer_intensity: DisplayIntensity,
}

Frame buffer structure

Fields
current: ColumnMajorBuffer

Actual, known, content of the display

buffer: ColumnMajorBuffer

Buffered content of the display, applied at the next call to render

column_index: usize

Current column index, used for the push (column) method

current_intensity: DisplayIntensity

Actual, known intensity of the display

buffer_intensity: DisplayIntensity

Buffered intensity of the display, applied at the next call to render

Implementations

impl FrameBuffer

fn render( &mut self, interface: &mut MAX7219<PinConnector<Pin<Output, PB3>, Pin<Output, PB2>, Pin<Output, PB5>>> )

Render buffered values (if necessary) to the display. If buffered and actual values are equal, do nothing.

fn push(&mut self, value: &[u8])

Push columns (as slice of u8) in the buffer, beginning at the current column index.

fn set_at(&mut self, column_index: usize, value: &[u8])

Set columns at the specified index, setting the column index after the newly set columns.

fn clear(&mut self)

Clear the buffer.