Struct alarm_clock::outputs::led_strip::LedStrip

pub struct LedStrip<const N: usize> {
    data_out: Pin<Output, PB0>,
    current: Color,
    buffer: Color,
}

Main structure for a LED strip with N LEDS

Fields
data_out: Pin<Output, PB0>

Pin connected to the digital input of the strip

current: Color

Current (actual) color

buffer: Color

Color to be set during the next rendering

Implementations

impl<const N: usize> LedStrip<N>

pub fn init<MODE: PinMode + Io>(data_out: Pin<MODE, PB0>) -> Self

Initialize the structure.

pub fn set_color(&mut self, color: Color)

Set a specific color to the LED strip

pub fn render(&mut self)

Actually render (set) the buffered color on the strip if different from the actual one.

fn write_color(&mut self)

Write the same color to all LEDS

fn write_bit(&mut self, bit: bool)

Write a single bit from the color. A bit true or false is decided by the duty cycle of the electrical raising/falling edges, hence the calls to the delays.

fn delay()

Generate a delay matching the specifications for a bit of the LED strip.