Struct alarm_clock::outputs::Outputs

pub struct Outputs {
    pub display: Display,
    pub led_strip: LedStrip<LED_STRIP_COUNT>,
    pub buzzer: Buzzer,
}

Fields
display: Display

Matrix-leds display, to print the time and some information

led_strip: LedStrip<LED_STRIP_COUNT>

Led strip to simulate dawn

buzzer: Buzzer

Buzzer alarm

Implementations

impl Outputs

pub fn init( data: Pin<impl Io, PB3>, cs: Pin<impl Io, PB2>, clk: Pin<impl Io, PB5>, led_strip: Pin<impl Io, PB0>, buzzer: Pin<impl Io, PD7> ) -> Self

Initialize the structure, in particular with the pinout. The matrix display uses a SPI interface, the other outputs use a simple digital pin.

pub fn render(&mut self)

Render the output. If the actual state is already the expected one, these functions do nothing (to spare computation time). These functions exist to avoid the cases when several updates of the same output occur during one loop of the main structure: the update might be a bit costly. The rendering is postponed until this function is called.