Struct alarm_clock::serial_commands::SerialBuffer

pub struct SerialBuffer<const WRITE_BUFFER_SIZE: usize, const READ_BUFFER_SIZE: usize> {
    output: ConstGenericRingBuffer<u8, WRITE_BUFFER_SIZE>,
    input: ConstGenericRingBuffer<u8, READ_BUFFER_SIZE>,
}

Buffers for serial communication, instantiated within a structure/a function rather than a global, mutex-protected variable.

Fields
output: ConstGenericRingBuffer<u8, WRITE_BUFFER_SIZE>input: ConstGenericRingBuffer<u8, READ_BUFFER_SIZE>
Implementations

impl<const WRITE_BUFFER_SIZE: usize, const READ_BUFFER_SIZE: usize> SerialBuffer<WRITE_BUFFER_SIZE, READ_BUFFER_SIZE>

pub fn flush(&mut self)

Actually flush (print) data from the output buffer on the USART.

pub fn load(&mut self)

Load bytes from the USART to the (input) ring buffer of this structure.

pub fn dequeue_command(&mut self) -> Result<Option<Command>, ()>

Try to dequeue a command from the input buffer. Return

Ok(Some()) when a command is identified

Err(()) when a \n separator was found, but no valid command could be parsed

Ok(None) most of the time, when there is no/not enough data Bytes are also removed from the ring buffer.

fn dequeue_to_array<const T: usize>(&mut self) -> [u8; T]

Dequeue from the input buffer to an array of a given size, dequeing (and dropping) the separator \n

fn decode_two_ascii_digits(d1: u8, d2: u8, base: u8) -> Result<u8, ()>

Decode d1d2 where d1 and d2 are digits in ascii, in a given base

fn decode_ascii_digit(d: u8) -> Result<u8, ()>

Map an ascii char to its value, expressed as as u8