Struct alarm_clock::MainState

pub(crate) struct MainState<I2C, const WRITE_BUFFER_SIZE: usize, const READ_BUFFER_SIZE: usize>where
    I2C: Write + WriteRead,{
    pub(crate) clocks: Clock<I2C>,
    pub(crate) inputs: Inputs,
    pub(crate) outputs: Outputs,
    pub(crate) serial_buffer: SerialBuffer<WRITE_BUFFER_SIZE, READ_BUFFER_SIZE>,
}

The main state of the whole program, updated at every loop, holding the memory.

Fields
clocks: Clock<I2C>

Clocks and alarms

inputs: Inputs

Inputs, either from the environment (light, etc.) or from the user (proximity, etc.)

outputs: Outputs

Outputs for the user (display, buzzer, etc.)

serial_buffer: SerialBuffer<WRITE_BUFFER_SIZE, READ_BUFFER_SIZE>

Serial I/O

Implementations

impl<const WRITE_BUFFER_SIZE: usize, const READ_BUFFER_SIZE: usize> MainState<I2c, WRITE_BUFFER_SIZE, READ_BUFFER_SIZE>

pub(crate) fn run(&mut self)

Run all the tasks needed to update the state/inputs/outputs.

pub(crate) fn update_inputs(&mut self)

Update all the inputs (clock, env, user), meant to be called before processing.

pub(crate) fn update_outputs(&mut self)

Update all the outputs for the user.

pub(crate) fn process(&mut self)

Process every module

pub(crate) fn process_alarm(&mut self)

Ack the alarm, if the conditions are met

pub(crate) fn process_display(&mut self)

Process the LED display output, value and intensity

pub(crate) fn process_led_strip(&mut self)

Switch on/off the led strip

pub(crate) fn process_buzzer(&mut self)

Decide to buzz… or not

pub(crate) fn process_command(&mut self)

Process commands received on serial input, if any